Added some method documentation and did some cleanup and style tweaking.

+++ src/Contact.java
@@ -4,7 +4,7 @@ public class Contact {
 
     Contact(String name, long phone) {
         this.name = name;
-        this.phone = (long) phone;
+        this.phone = phone;
     }
 
     @Override
@@ -20,6 +20,7 @@ public class Contact {
         return phone;
     }
 
+    // converts the contact into a String formatted to look like "name,phonenumber"
     public String formatForStorage() {
         return String.format("%s,%s", name, phone);
     }
